+++ contracts/src/access/ownable.rs
@@ -74,9 +74,6 @@ pub struct Ownable {
 /// Interface for an [`Ownable`] contract.
 #[interface_id]
 pub trait IOwnable {
-    /// The error type associated to the trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Returns the address of the current owner.
     ///
     /// # Arguments
@@ -99,10 +96,7 @@ pub trait IOwnable {
     /// # Events
     ///
     /// * [`OwnershipTransferred`].
-    fn transfer_ownership(
-        &mut self,
-        new_owner: Address,
-    ) -> Result<(), <Self as IOwnable>::Error>;
+    fn transfer_ownership(&mut self, new_owner: Address) -> Result<(), Error>;
 
     /// Leaves the contract without owner. It will not be possible to call
     /// functions that require `only_owner`. Can only be called by the current
@@ -122,21 +116,16 @@ pub trait IOwnable {
     /// # Events
     ///
     /// * [`OwnershipTransferred`].
-    fn renounce_ownership(&mut self) -> Result<(), <Self as IOwnable>::Error>;
+    fn renounce_ownership(&mut self) -> Result<(), Error>;
 }
 
 #[public]
 impl IOwnable for Ownable {
-    type Error = Error;
-
     fn owner(&self) -> Address {
         self.owner.get()
     }
 
-    fn transfer_ownership(
-        &mut self,
-        new_owner: Address,
-    ) -> Result<(), <Self as IOwnable>::Error> {
+    fn transfer_ownership(&mut self, new_owner: Address) -> Result<(), Error> {
         self.only_owner()?;
 
         if new_owner.is_zero() {
@@ -150,7 +139,7 @@ impl IOwnable for Ownable {
         Ok(())
     }
 
-    fn renounce_ownership(&mut self) -> Result<(), <Self as IOwnable>::Error> {
+    fn renounce_ownership(&mut self) -> Result<(), Error> {
         self.only_owner()?;
         self._transfer_ownership(Address::ZERO);
         Ok(())