+++ contracts/src/access/ownable_two_step.rs
@@ -75,9 +75,6 @@ impl DerefMut for Ownable2Step {
 /// Interface for an [`Ownable2Step`] contract.
 #[interface_id]
 pub trait IOwnable2Step {
-    /// The error type associated to the trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Returns the address of the current owner.
     ///
     /// Re-export of [`Ownable::owner`].
@@ -117,7 +114,7 @@ pub trait IOwnable2Step {
     fn transfer_ownership(
         &mut self,
         new_owner: Address,
-    ) -> Result<(), <Self as IOwnable2Step>::Error>;
+    ) -> Result<(), ownable::Error>;
 
     /// Accepts the ownership of the contract.
     /// Can only be called by the pending owner.
@@ -134,9 +131,7 @@ pub trait IOwnable2Step {
     /// # Events
     ///
     /// * [`crate::access::ownable::OwnershipTransferred`].
-    fn accept_ownership(
-        &mut self,
-    ) -> Result<(), <Self as IOwnable2Step>::Error>;
+    fn accept_ownership(&mut self) -> Result<(), ownable::Error>;
 
     /// Leaves the contract without owner. It will not be possible to call
     /// [`Ownable::only_owner`] functions. Can only be called by the current
@@ -156,15 +151,11 @@ pub trait IOwnable2Step {
     /// # Events
     ///
     /// * [`crate::access::ownable::OwnershipTransferred`].
-    fn renounce_ownership(
-        &mut self,
-    ) -> Result<(), <Self as IOwnable2Step>::Error>;
+    fn renounce_ownership(&mut self) -> Result<(), ownable::Error>;
 }
 
 #[public]
 impl IOwnable2Step for Ownable2Step {
-    type Error = ownable::Error;
-
     fn owner(&self) -> Address {
         self.ownable.owner()
     }
@@ -176,7 +167,7 @@ impl IOwnable2Step for Ownable2Step {
     fn transfer_ownership(
         &mut self,
         new_owner: Address,
-    ) -> Result<(), <Self as IOwnable2Step>::Error> {
+    ) -> Result<(), ownable::Error> {
         self.ownable.only_owner()?;
         self.pending_owner.set(new_owner);
 
@@ -188,9 +179,7 @@ impl IOwnable2Step for Ownable2Step {
         Ok(())
     }
 
-    fn accept_ownership(
-        &mut self,
-    ) -> Result<(), <Self as IOwnable2Step>::Error> {
+    fn accept_ownership(&mut self) -> Result<(), ownable::Error> {
         let sender = msg::sender();
         let pending_owner = self.pending_owner();
         if sender != pending_owner {
@@ -202,9 +191,7 @@ impl IOwnable2Step for Ownable2Step {
         Ok(())
     }
 
-    fn renounce_ownership(
-        &mut self,
-    ) -> Result<(), <Self as IOwnable2Step>::Error> {
+    fn renounce_ownership(&mut self) -> Result<(), ownable::Error> {
         self.ownable.only_owner()?;
         self._transfer_ownership(Address::ZERO);
         Ok(())