+++ contracts/src/finance/vesting_wallet.rs
@@ -188,9 +188,6 @@ unsafe impl TopLevelStorage for VestingWallet {}
 /// Required interface of a [`VestingWallet`] compliant contract.
 #[interface_id]
 pub trait IVestingWallet {
-    /// The error type associated to this trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Returns the address of the current owner.
     ///
     /// Re-export of [`Ownable::owner`].
@@ -220,10 +217,7 @@ pub trait IVestingWallet {
     /// # Events
     ///
     /// * [`ownable::OwnershipTransferred`].
-    fn transfer_ownership(
-        &mut self,
-        new_owner: Address,
-    ) -> Result<(), <Self as IVestingWallet>::Error>;
+    fn transfer_ownership(&mut self, new_owner: Address) -> Result<(), Error>;
 
     /// Leaves the contract without owner. It will not be possible to call
     /// [`Ownable::only_owner`] functions. Can only be called by the current
@@ -245,9 +239,7 @@ pub trait IVestingWallet {
     /// # Events
     ///
     /// * [`ownable::OwnershipTransferred`].
-    fn renounce_ownership(
-        &mut self,
-    ) -> Result<(), <Self as IVestingWallet>::Error>;
+    fn renounce_ownership(&mut self) -> Result<(), Error>;
 
     /// Getter for the start timestamp.
     ///
@@ -317,10 +309,7 @@ pub trait IVestingWallet {
     /// * If total allocation exceeds `U256::MAX`.
     /// * If scaled, total allocation (mid calculation) exceeds `U256::MAX`.
     #[selector(name = "releasable")]
-    fn releasable_erc20(
-        &mut self,
-        token: Address,
-    ) -> Result<U256, <Self as IVestingWallet>::Error>;
+    fn releasable_erc20(&mut self, token: Address) -> Result<U256, Error>;
 
     /// Release the native tokens (Ether) that have already vested.
     ///
@@ -341,7 +330,7 @@ pub trait IVestingWallet {
     /// * If total allocation exceeds `U256::MAX`.
     /// * If scaled total allocation (mid calculation) exceeds `U256::MAX`.
     #[selector(name = "release")]
-    fn release_eth(&mut self) -> Result<(), <Self as IVestingWallet>::Error>;
+    fn release_eth(&mut self) -> Result<(), Error>;
 
     /// Release the tokens that have already vested.
     ///
@@ -365,10 +354,7 @@ pub trait IVestingWallet {
     /// * If total allocation exceeds `U256::MAX`.
     /// * If scaled, total allocation (mid calculation) exceeds `U256::MAX`.
     #[selector(name = "release")]
-    fn release_erc20(
-        &mut self,
-        token: Address,
-    ) -> Result<(), <Self as IVestingWallet>::Error>;
+    fn release_erc20(&mut self, token: Address) -> Result<(), Error>;
 
     /// Calculates the amount of Ether that has already vested.
     /// The Default implementation is a linear vesting curve.
@@ -407,27 +393,20 @@ pub trait IVestingWallet {
         &mut self,
         token: Address,
         timestamp: u64,
-    ) -> Result<U256, <Self as IVestingWallet>::Error>;
+    ) -> Result<U256, Error>;
 }
 
 #[public]
 impl IVestingWallet for VestingWallet {
-    type Error = Error;
-
     fn owner(&self) -> Address {
         self.ownable.owner()
     }
 
-    fn transfer_ownership(
-        &mut self,
-        new_owner: Address,
-    ) -> Result<(), <Self as IVestingWallet>::Error> {
+    fn transfer_ownership(&mut self, new_owner: Address) -> Result<(), Error> {
         Ok(self.ownable.transfer_ownership(new_owner)?)
     }
 
-    fn renounce_ownership(
-        &mut self,
-    ) -> Result<(), <Self as IVestingWallet>::Error> {
+    fn renounce_ownership(&mut self) -> Result<(), Error> {
         Ok(self.ownable.renounce_ownership()?)
     }
 
@@ -463,10 +442,7 @@ impl IVestingWallet for VestingWallet {
     }
 
     #[selector(name = "releasable")]
-    fn releasable_erc20(
-        &mut self,
-        token: Address,
-    ) -> Result<U256, <Self as IVestingWallet>::Error> {
+    fn releasable_erc20(&mut self, token: Address) -> Result<U256, Error> {
         let vested = self.vested_amount_erc20(token, block::timestamp())?;
         // SAFETY: total vested amount is by definition greater than or equal to
         // the released amount.
@@ -474,7 +450,7 @@ impl IVestingWallet for VestingWallet {
     }
 
     #[selector(name = "release")]
-    fn release_eth(&mut self) -> Result<(), <Self as IVestingWallet>::Error> {
+    fn release_eth(&mut self) -> Result<(), Error> {
         let amount = self.releasable_eth();
 
         self.released.add_assign_checked(
@@ -492,10 +468,7 @@ impl IVestingWallet for VestingWallet {
     }
 
     #[selector(name = "release")]
-    fn release_erc20(
-        &mut self,
-        token: Address,
-    ) -> Result<(), <Self as IVestingWallet>::Error> {
+    fn release_erc20(&mut self, token: Address) -> Result<(), Error> {
         let amount = self.releasable_erc20(token)?;
         let owner = self.ownable.owner();
 
@@ -525,7 +498,7 @@ impl IVestingWallet for VestingWallet {
         &mut self,
         token: Address,
         timestamp: u64,
-    ) -> Result<U256, <Self as IVestingWallet>::Error> {
+    ) -> Result<U256, Error> {
         let erc20 = Erc20Interface::new(token);
         let balance = erc20
             .balance_of(Call::new_in(self), contract::address())