+++ contracts/src/lib.rs
@@ -38,11 +38,7 @@ impl MyContract { }
 ```
 */
 
-#![allow(
-    clippy::module_name_repetitions,
-    clippy::used_underscore_items,
-    deprecated
-)]
+#![allow(clippy::module_name_repetitions, deprecated)]
 #![cfg_attr(not(feature = "std"), no_std, no_main)]
 #![cfg_attr(coverage_nightly, feature(coverage_attribute))]
 #![deny(rustdoc::broken_intra_doc_links)]