+++ contracts/src/token/erc1155/extensions/metadata_uri.rs
@@ -100,7 +100,8 @@ mod tests {
 
     #[motsu::test]
     fn interface_id() {
-        let actual = <Erc1155MetadataUri as IErc1155MetadataUri>::interface_id();
+        let actual =
+            <Erc1155MetadataUri as IErc1155MetadataUri>::interface_id();
         let expected = 0x0e89341c;
         assert_eq!(actual, expected);
     }