+++ contracts/src/token/erc1155/extensions/supply.rs
@@ -100,8 +100,6 @@ impl IErc1155Supply for Erc1155Supply {
 
 #[public]
 impl IErc1155 for Erc1155Supply {
-    type Error = erc1155::Error;
-
     fn balance_of(&self, account: Address, id: U256) -> U256 {
         self.erc1155.balance_of(account, id)
     }
@@ -110,7 +108,7 @@ impl IErc1155 for Erc1155Supply {
         &self,
         accounts: Vec<Address>,
         ids: Vec<U256>,
-    ) -> Result<Vec<U256>, <Self as IErc1155>::Error> {
+    ) -> Result<Vec<U256>, erc1155::Error> {
         self.erc1155.balance_of_batch(accounts, ids)
     }
 
@@ -118,7 +116,7 @@ impl IErc1155 for Erc1155Supply {
         &mut self,
         operator: Address,
         approved: bool,
-    ) -> Result<(), <Self as IErc1155>::Error> {
+    ) -> Result<(), erc1155::Error> {
         self.erc1155.set_approval_for_all(operator, approved)
     }
 
@@ -133,7 +131,7 @@ impl IErc1155 for Erc1155Supply {
         id: U256,
         value: U256,
         data: Bytes,
-    ) -> Result<(), <Self as IErc1155>::Error> {
+    ) -> Result<(), erc1155::Error> {
         self.erc1155.authorize_transfer(from)?;
         self.do_safe_transfer_from(from, to, vec![id], vec![value], &data)
     }
@@ -145,7 +143,7 @@ impl IErc1155 for Erc1155Supply {
         ids: Vec<U256>,
         values: Vec<U256>,
         data: Bytes,
-    ) -> Result<(), <Self as IErc1155>::Error> {
+    ) -> Result<(), erc1155::Error> {
         self.erc1155.authorize_transfer(from)?;
         self.do_safe_transfer_from(from, to, ids, values, &data)
     }