+++ contracts/src/token/erc1155/mod.rs
@@ -218,9 +218,6 @@ unsafe impl TopLevelStorage for Erc1155 {}
 /// Required interface of an [`Erc1155`] compliant contract.
 #[interface_id]
 pub trait IErc1155: IErc165 {
-    /// The error type associated to this ERC-1155 trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Returns the value of tokens of type `id` owned by `account`.
     ///
     /// # Arguments
@@ -246,7 +243,7 @@ pub trait IErc1155: IErc165 {
         &self,
         accounts: Vec<Address>,
         ids: Vec<U256>,
-    ) -> Result<Vec<U256>, <Self as IErc1155>::Error>;
+    ) -> Result<Vec<U256>, Error>;
 
     /// Grants or revokes permission to `operator`
     /// to transfer the caller's tokens, according to `approved`.
@@ -270,7 +267,7 @@ pub trait IErc1155: IErc165 {
         &mut self,
         operator: Address,
         approved: bool,
-    ) -> Result<(), <Self as IErc1155>::Error>;
+    ) -> Result<(), Error>;
 
     /// Returns true if `operator` is approved to transfer `account`'s
     /// tokens.
@@ -318,7 +315,7 @@ pub trait IErc1155: IErc165 {
         id: U256,
         value: U256,
         data: Bytes,
-    ) -> Result<(), <Self as IErc1155>::Error>;
+    ) -> Result<(), Error>;
 
     /// Batched version of [`IErc1155::safe_transfer_from`].
     ///
@@ -358,13 +355,11 @@ pub trait IErc1155: IErc165 {
         ids: Vec<U256>,
         values: Vec<U256>,
         data: Bytes,
-    ) -> Result<(), <Self as IErc1155>::Error>;
+    ) -> Result<(), Error>;
 }
 
 #[public]
 impl IErc1155 for Erc1155 {
-    type Error = Error;
-
     fn balance_of(&self, account: Address, id: U256) -> U256 {
         self.balances.get(id).get(account)
     }
@@ -373,7 +368,7 @@ impl IErc1155 for Erc1155 {
         &self,
         accounts: Vec<Address>,
         ids: Vec<U256>,
-    ) -> Result<Vec<U256>, <Self as IErc1155>::Error> {
+    ) -> Result<Vec<U256>, Error> {
         Self::require_equal_arrays_length(&ids, &accounts)?;
 
         let balances: Vec<U256> = accounts
@@ -389,7 +384,7 @@ impl IErc1155 for Erc1155 {
         &mut self,
         operator: Address,
         approved: bool,
-    ) -> Result<(), <Self as IErc1155>::Error> {
+    ) -> Result<(), Error> {
         self._set_approval_for_all(msg::sender(), operator, approved)
     }
 
@@ -404,7 +399,7 @@ impl IErc1155 for Erc1155 {
         id: U256,
         value: U256,
         data: Bytes,
-    ) -> Result<(), <Self as IErc1155>::Error> {
+    ) -> Result<(), Error> {
         self.authorize_transfer(from)?;
         self.do_safe_transfer_from(from, to, vec![id], vec![value], &data)
     }
@@ -416,7 +411,7 @@ impl IErc1155 for Erc1155 {
         ids: Vec<U256>,
         values: Vec<U256>,
         data: Bytes,
-    ) -> Result<(), <Self as IErc1155>::Error> {
+    ) -> Result<(), Error> {
         self.authorize_transfer(from)?;
         self.do_safe_transfer_from(from, to, ids, values, &data)
     }