+++ contracts/src/token/erc20/extensions/burnable.rs
@@ -9,9 +9,6 @@ use crate::token::erc20::{self, Erc20};
 /// their own tokens and those that they have an allowance for,
 /// in a way that can be recognized off-chain (via event analysis).
 pub trait IErc20Burnable {
-    /// The error type associated to this ERC-20 Burnable trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Destroys a `value` amount of tokens from the caller, lowering the total
     /// supply.
     ///
@@ -27,10 +24,7 @@ pub trait IErc20Burnable {
     /// # Events
     ///
     /// * [`erc20::Transfer`].
-    fn burn(
-        &mut self,
-        value: U256,
-    ) -> Result<(), <Self as IErc20Burnable>::Error>;
+    fn burn(&mut self, value: U256) -> Result<(), erc20::Error>;
 
     /// Destroys a `value` amount of tokens from `account`, lowering the total
     /// supply.
@@ -56,16 +50,11 @@ pub trait IErc20Burnable {
         &mut self,
         account: Address,
         value: U256,
-    ) -> Result<(), <Self as IErc20Burnable>::Error>;
+    ) -> Result<(), erc20::Error>;
 }
 
 impl IErc20Burnable for Erc20 {
-    type Error = erc20::Error;
-
-    fn burn(
-        &mut self,
-        value: U256,
-    ) -> Result<(), <Self as IErc20Burnable>::Error> {
+    fn burn(&mut self, value: U256) -> Result<(), erc20::Error> {
         self._burn(msg::sender(), value)
     }
 
@@ -73,7 +62,7 @@ impl IErc20Burnable for Erc20 {
         &mut self,
         account: Address,
         value: U256,
-    ) -> Result<(), <Self as IErc20Burnable>::Error> {
+    ) -> Result<(), erc20::Error> {
         self._spend_allowance(account, msg::sender(), value)?;
         self._burn(account, value)
     }