+++ contracts/src/token/erc20/mod.rs
@@ -141,9 +141,6 @@ pub struct Erc20 {
 /// Required interface of an [`Erc20`] compliant contract.
 #[interface_id]
 pub trait IErc20 {
-    /// The error type associated to this ERC-20 trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Returns the number of tokens in existence.
     ///
     /// # Arguments
@@ -178,11 +175,7 @@ pub trait IErc20 {
     /// # Events
     ///
     /// * [`Transfer`].
-    fn transfer(
-        &mut self,
-        to: Address,
-        value: U256,
-    ) -> Result<bool, <Self as IErc20>::Error>;
+    fn transfer(&mut self, to: Address, value: U256) -> Result<bool, Error>;
 
     /// Returns the remaining number of tokens that `spender` will be allowed
     /// to spend on behalf of `owner` through `transfer_from`. This is zero by
@@ -223,11 +216,8 @@ pub trait IErc20 {
     /// # Events
     ///
     /// * [`Approval`].
-    fn approve(
-        &mut self,
-        spender: Address,
-        value: U256,
-    ) -> Result<bool, <Self as IErc20>::Error>;
+    fn approve(&mut self, spender: Address, value: U256)
+        -> Result<bool, Error>;
 
     /// Moves a `value` number of tokens from `from` to `to` using the
     /// allowance mechanism. `value` is then deducted from the caller's
@@ -264,13 +254,11 @@ pub trait IErc20 {
         from: Address,
         to: Address,
         value: U256,
-    ) -> Result<bool, <Self as IErc20>::Error>;
+    ) -> Result<bool, Error>;
 }
 
 #[public]
 impl IErc20 for Erc20 {
-    type Error = Error;
-
     fn total_supply(&self) -> U256 {
         self.total_supply.get()
     }
@@ -279,11 +267,7 @@ impl IErc20 for Erc20 {
         self.balances.get(account)
     }
 
-    fn transfer(
-        &mut self,
-        to: Address,
-        value: U256,
-    ) -> Result<bool, <Self as IErc20>::Error> {
+    fn transfer(&mut self, to: Address, value: U256) -> Result<bool, Error> {
         let from = msg::sender();
         self._transfer(from, to, value)?;
         Ok(true)
@@ -297,7 +281,7 @@ impl IErc20 for Erc20 {
         &mut self,
         spender: Address,
         value: U256,
-    ) -> Result<bool, <Self as IErc20>::Error> {
+    ) -> Result<bool, Error> {
         let owner = msg::sender();
         self._approve(owner, spender, value, true)
     }
@@ -307,7 +291,7 @@ impl IErc20 for Erc20 {
         from: Address,
         to: Address,
         value: U256,
-    ) -> Result<bool, <Self as IErc20>::Error> {
+    ) -> Result<bool, Error> {
         let spender = msg::sender();
         self._spend_allowance(from, spender, value)?;
         self._transfer(from, to, value)?;