+++ contracts/src/token/erc20/utils/safe_erc20.rs
@@ -96,9 +96,6 @@ unsafe impl TopLevelStorage for SafeErc20 {}
 /// Required interface of a [`SafeErc20`] utility contract.
 #[interface_id]
 pub trait ISafeErc20 {
-    /// The error type associated to this trait implementation.
-    type Error: Into<alloc::vec::Vec<u8>>;
-
     /// Transfer `value` amount of `token` from the calling contract to `to`. If
     /// `token` returns no value, non-reverting calls are assumed to be
     /// successful.
@@ -120,7 +117,7 @@ pub trait ISafeErc20 {
         token: Address,
         to: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error>;
+    ) -> Result<(), Error>;
 
     /// Transfer `value` amount of `token` from `from` to `to`, spending the
     /// approval given by `from` to the calling contract. If `token` returns
@@ -145,7 +142,7 @@ pub trait ISafeErc20 {
         from: Address,
         to: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error>;
+    ) -> Result<(), Error>;
 
     /// Increase the calling contract's allowance toward `spender` by `value`.
     /// If `token` returns no value, non-reverting calls are assumed to be
@@ -172,7 +169,7 @@ pub trait ISafeErc20 {
         token: Address,
         spender: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error>;
+    ) -> Result<(), Error>;
 
     /// Decrease the calling contract's allowance toward `spender` by
     /// `requested_decrease`. If `token` returns no value, non-reverting
@@ -197,7 +194,7 @@ pub trait ISafeErc20 {
         token: Address,
         spender: Address,
         requested_decrease: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error>;
+    ) -> Result<(), Error>;
 
     /// Set the calling contract's allowance toward `spender` to `value`. If
     /// `token` returns no value, non-reverting calls are assumed to be
@@ -221,19 +218,17 @@ pub trait ISafeErc20 {
         token: Address,
         spender: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error>;
+    ) -> Result<(), Error>;
 }
 
 #[public]
 impl ISafeErc20 for SafeErc20 {
-    type Error = Error;
-
     fn safe_transfer(
         &mut self,
         token: Address,
         to: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error> {
+    ) -> Result<(), Error> {
         let call = IErc20::transferCall { to, value };
 
         Self::call_optional_return(token, &call)
@@ -245,7 +240,7 @@ impl ISafeErc20 for SafeErc20 {
         from: Address,
         to: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error> {
+    ) -> Result<(), Error> {
         let call = IErc20::transferFromCall { from, to, value };
 
         Self::call_optional_return(token, &call)
@@ -256,7 +251,7 @@ impl ISafeErc20 for SafeErc20 {
         token: Address,
         spender: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error> {
+    ) -> Result<(), Error> {
         let current_allowance = Self::allowance(token, spender)?;
         let new_allowance = current_allowance
             .checked_add(value)
@@ -269,7 +264,7 @@ impl ISafeErc20 for SafeErc20 {
         token: Address,
         spender: Address,
         requested_decrease: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error> {
+    ) -> Result<(), Error> {
         let current_allowance = Self::allowance(token, spender)?;
 
         if current_allowance < requested_decrease {
@@ -293,7 +288,7 @@ impl ISafeErc20 for SafeErc20 {
         token: Address,
         spender: Address,
         value: U256,
-    ) -> Result<(), <Self as ISafeErc20>::Error> {
+    ) -> Result<(), Error> {
         let approve_call = IErc20::approveCall { spender, value };
 
         // Try performing the approval with the desired value.